issues: Simpler implementation of import path to absolute dir resolution.
I needed this functionality quite often in many packages, so over time
I've come down to this solution. It uses go/build.Import with build.FindOnly
flag. Given Go 1 promise, I think it's safe to rely on that behavior
and there's no need to reimplement it from scratch.

It seems to be functionally equivalent to the previous code, but much
shorter, simpler and easier to maintain.