cmd/servegoissues: Display link to view original issue.
Previously, it would display a "Sign in to comment" message which was
not applicable to this view-only Go issue viewer. User logins are not
supported here. Even if they were, it wouldn't make sense to post new
comments to a mirror of Go issues.

Also increase bottom margin to make it more clear when at bottom of
the issue.

Resolves #9.