cmd/servegoissues: Disable reactions.
The new issuesapp API (see shurcooL/issuesapp#1) requires the user to
provide emojis (they are now served locally rather than from the
internet), unless DisableReactions is true.

Given that the Go issues data does not include any reactions, and it's
not possible to login to react, it seems to make more sense to outright
remove them.